reorganize to remove circular import
