+++ sr-olthad/src/sr_olthad/lm_step.py
@@ -10,9 +10,10 @@ from sr_olthad.common.agents import (
     LmRetryHandler,
 )
 from sr_olthad.common.schema import InstructLmMessage
+from sr_olthad.common.utils import render_single_turn_prompt_templates_and_get_messages
 from sr_olthad.olthad import PendingOlthadUpdate
-from sr_olthad.schema import CommonUserPromptInputData, LmAgentName
-from sr_olthad.utils import get_input_messages
+from sr_olthad.registry import LM_AGENT_CONFIGS_REGISTRY, PROMPT_REGISTRIES_REGISTRY
+from sr_olthad.schema import CommonSysPromptInputData, CommonUserPromptInputData, LmAgentName
 
 
 class PreLmStepEmission(BaseModel):
@@ -130,3 +131,31 @@ class LmStepTemplate:
                     return return_if_approved
 
         return _lm_step
+
+
+def get_input_messages(
+    lm_agent_name: LmAgentName,
+    user_prompt_input_data: CommonUserPromptInputData,
+    get_domain_specific_insert: GetDomainSpecificInsert | None = None,
+) -> list[InstructLmMessage]:
+    """
+    Gets agent prompt templates from the registries and renders necessary data into them.
+    """
+    cfg = LM_AGENT_CONFIGS_REGISTRY[lm_agent_name]
+    prompt_registry = PROMPT_REGISTRIES_REGISTRY[lm_agent_name]
+
+    if get_domain_specific_insert is None:
+        sys_prompt_input_data = None
+    else:
+        sys_prompt_input_data = CommonSysPromptInputData(
+            domain_specific_insert=get_domain_specific_insert(
+                lm_agent_name, user_prompt_input_data
+            ),
+        )
+
+    return render_single_turn_prompt_templates_and_get_messages(
+        user_prompt_template=prompt_registry[cfg.PROMPTS_VERSION].user_prompt_template,
+        user_prompt_input_data=user_prompt_input_data,
+        sys_prompt_template=prompt_registry[cfg.PROMPTS_VERSION].sys_prompt_template,
+        sys_prompt_input_data=sys_prompt_input_data,
+    )
